Crate rhdl_vlog

Re-exports

pub use atoms::BitRange;

pub use atoms::DeclKind;

pub use atoms::Declaration;

pub use atoms::DeclarationList;

pub use atoms::Direction;

pub use atoms::HDLKind;

pub use atoms::Port;

pub use atoms::SignedWidth;

pub use atoms::WidthSpec;

pub use crate::atoms::LitVerilog;

pub use crate::expr::Expr;

pub use crate::expr::ExprConcat;

pub use crate::expr::ExprDynIndex;

pub use crate::expr::ExprIndex;

pub use crate::formatter::Formatter;

pub use crate::formatter::Pretty;

pub use crate::kw_ops::BinaryOp;

pub use crate::kw_ops::DynOp;

pub use crate::kw_ops::UnaryOp;

pub use crate::stmt::Stmt;

pub use builder::*;

Modules

atoms

builder

Builder functions for constructing Verilog AST nodes

expr

formatter

kw_ops

stmt

Macros

parse_quote_miette

Structs

FunctionDef

Initial

Item

ItemList

ModuleDef

ModuleList

ParseError

SynthesisAttribute

SynthesisAttributeList

Enums

Attribute

ItemKind

Functions

format_verilog_for_error_display

Format Verilog code for better error display by adding line breaks without requiring the code to be syntactically valid.